Move handling of layout manager from qgisapp -> QgsAppWindowManager
