+++ src/app/qgsappwindowmanager.h
@@ -21,6 +21,7 @@
 #include <QPointer>
 
 class QgsStyleManagerDialog;
+class QgsLayoutManagerDialog;
 
 /**
  * \ingroup gui
@@ -30,13 +31,31 @@ class QgsAppWindowManager : public QgsWindowManagerInterface
 {
   public:
 
+    //! Application-only QGIS dialogs
+    enum ApplicationDialog
+    {
+      DialogLayoutManager = 0, //!< Layout manager dialog
+    };
+
     QgsAppWindowManager() = default;
     ~QgsAppWindowManager();
 
     QWidget *openStandardDialog( QgsWindowManagerInterface::StandardDialog dialog ) override;
 
+    /**
+     * Opens an instance of a application QGIS dialog. Depending on the dialog,
+     * this may either open a new instance of the dialog or bring an
+     * existing instance to the foreground.
+     *
+     * Returns the dialog if shown, or nullptr if the dialog either could not be
+     * created.
+     */
+    QWidget *openApplicationDialog( ApplicationDialog dialog );
+
   private:
     QPointer< QgsStyleManagerDialog > mStyleManagerDialog;
+    QPointer< QgsLayoutManagerDialog > mLayoutManagerDialog;
+
 };
